+++ Changelog.md
@@ -3,6 +3,8 @@
 ### Added
 
 -   Reconnecting status in device selector.
+-   Custom control to set net core upload delay
+-   Persist net core upload delay for each device serial number
 
 ### Fixed